improve body detection
